Refactor: create utils folder and improve code organization
- Create utils folder with reusable functions:
  - github-api.ts: GitHub API headers utility
  - repository.ts: Repository name mapping and fallback languages
  - validation.ts: Form validation and data extraction utilities

- Update api/github/route.ts to use shared GitHub headers
- Enhance contact-actions.tsx with proper email validation
- Clean up cypress.production.config.ts comments
- Fix import paths in projects/page.tsx and cv/page.tsx
